Made some adjustments

+++ migrations/1643425749__setup_desktop.sh
@@ -28,6 +28,7 @@ if [ "$ENVIRONMENT" != "sandbox" ]; then
     mpc \
     mpd \
     mpv \
+    ncmpcpp \
     network-manager-applet \
     p7zip \
     pass \
@@ -62,46 +63,82 @@ if [ "$ENVIRONMENT" != "sandbox" ]; then
   # Enable bluetooth
   sudo systemctl enable --now bluetooth.service
 
+  # Enable pulseaudio
+  systemctl --user enable --now pulseaudio
+
   # icons
-  ln -sf "$_DOTFILES_HOME/icons" "$HOME/.icons"
+  if [[ ! -e "$_DOTFILES_HOME/icons" ]]; then
+    ln -sf "$_DOTFILES_HOME/icons" "$HOME/.icons"
+  fi
 
   # themes
-  ln -sf "$_DOTFILES_HOME/gtk-2.0" "$XDG_CONFIG_HOME/gtk-2.0"
-  ln -sf "$_DOTFILES_HOME/gtk-3.0" "$XDG_CONFIG_HOME/gtk-3.0"
+  if [[ ! -e "$_DOTFILES_HOME/gtk-2.0" ]]; then
+    ln -sf "$_DOTFILES_HOME/gtk-2.0" "$XDG_CONFIG_HOME/gtk-2.0"
+  fi
+  if [[ ! -e "$_DOTFILES_HOME/gtk-3.0" ]]; then
+    ln -sf "$_DOTFILES_HOME/gtk-3.0" "$XDG_CONFIG_HOME/gtk-3.0"
+  fi
 
   # X11
-  ln -sf "$_DOTFILES_HOME/X11" "$XDG_CONFIG_HOME/X11"
+  if [[ ! -e "$_DOTFILES_HOME/X11" ]]; then
+    ln -sf "$_DOTFILES_HOME/X11" "$XDG_CONFIG_HOME/X11"
+  fi
+
+  # xmobar
+  if [[ ! -e "$_DOTFILES_HOME/xmobar" ]]; then
+    ln -sf "$_DOTFILES_HOME/xmobar" "$XDG_CONFIG_HOME/xmobar"
+  fi
+  (
+    cd "$XDG_CONFIG_HOME/xmobar"
+    cabal install --overwrite-policy=always
+  )
 
   # xmonad
-  ln -sf "$_DOTFILES_HOME/xmonad" "$XDG_CONFIG_HOME/xmonad"
+  if [[ ! -e "$_DOTFILES_HOME/xmonad" ]]; then
+    ln -sf "$_DOTFILES_HOME/xmonad" "$XDG_CONFIG_HOME/xmonad"
+  fi
   (
     cd "$XDG_CONFIG_HOME/xmonad"
-    cabal install
+    cabal install --overwrite-policy=always
     xmonad --recompile
   )
 
   # picom
-  ln -sf "$_DOTFILES_HOME/picom" "$XDG_CONFIG_HOME/picom"
+  if [[ ! -e "$_DOTFILES_HOME/picom/picom.conf" ]]; then
+    ln -sf "$_DOTFILES_HOME/picom/picom.conf" "$XDG_CONFIG_HOME/picom/picom.conf"
+  fi
 
   # stalonetray
-  ln -sf "$_DOTFILES_HOME/stalonetray" "$XDG_CONFIG_HOME/stalonetray"
+  if [[ ! -e "$_DOTFILES_HOME/stalonetray" ]]; then
+    ln -sf "$_DOTFILES_HOME/stalonetray" "$XDG_CONFIG_HOME/stalonetray"
+  fi
 
   # alacritty
-  ln -sf "$_DOTFILES_HOME/alacritty" "$XDG_CONFIG_HOME/alacritty"
+  if [[ ! -e "$_DOTFILES_HOME/alacritty" ]]; then
+    ln -sf "$_DOTFILES_HOME/alacritty" "$XDG_CONFIG_HOME/alacritty"
+  fi
   (
     cd "$XDG_CONFIG_HOME/alacritty"
     ./merge.sh
   )
 
   # xrandr
-  ln -sf "$_DOTFILES_HOME/xrandr" "$XDG_CONFIG_HOME/xrandr"
+  if [[ ! -e "$_DOTFILES_HOME/xrandr" ]]; then
+    ln -sf "$_DOTFILES_HOME/xrandr" "$XDG_CONFIG_HOME/xrandr"
+  fi
 
+  # TODO NEEDS FIXED!
   # mpd
-  ln -sf "$_DOTFILES_HOME/mpd" "$XDG_CONFIG_HOME/mpd"
+  if [[ ! -e "$_DOTFILES_HOME/mpd" ]]; then
+    ln -sf "$_DOTFILES_HOME/mpd" "$XDG_CONFIG_HOME/mpd"
+  fi
   mkdir -p "$XDG_DATA_HOME/mpd/playlists"
-  systemctl --user enable mpd
+  mkdir -p "$XDG_CACHE_HOME/mpd"
+  systemctl --user enable --now mpd
   mpc update
 
   # ncmpcpp
-  ln -sf "$_DOTFILES_HOME/ncmpcpp" "$XDG_CONFIG_HOME/ncmpcpp"
+  if [[ ! -e "$_DOTFILES_HOME/ncmpcpp" ]]; then
+    ln -sf "$_DOTFILES_HOME/ncmpcpp" "$XDG_CONFIG_HOME/ncmpcpp"
+  fi
 fi
